requirements.txt
discord.py==2.7.1
aiohttp==3.13.5
python-dotenv==1.2.2
pydantic-settings==2.13.1
langchain-core==1.2.26
langchain-community==0.4.1
langchain-google-genai==4.2.1
langchain-groq==1.1.2
google-genai==1.70.0
groq==0.37.1
duckduckgo-search==8.1.1
sentence-transformers==5.3.0
numpy==2.4.4
torch==2.11.0
